+++ agents/src/tokenize/tokenizer.test.ts
@@ -13,6 +13,8 @@ const TEXT =
   'This is a test. Another test. ' +
   'A short sentence. ' +
   'A longer sentence that is longer than the previous sentence. ' +
+  'Find additional resources on livekit.com. ' +
+  'Find additional resources on docs.livekit.com. ' +
   'f(x) = x * 2.54 + 42. ' +
   'Hey! Hi! Hello! ';
 
@@ -22,6 +24,8 @@ const EXPECTED_MIN_20 = [
   'Mr. Theo is testing the sentence tokenizer.',
   'This is a test. Another test.',
   'A short sentence. A longer sentence that is longer than the previous sentence.',
+  'Find additional resources on livekit.com.',
+  'Find additional resources on docs.livekit.com.',
   'f(x) = x * 2.54 + 42.',
   'Hey! Hi! Hello!',
 ];